updated few comments and notebook
