+++ kedro-databricks-bootcamp/01_intro/solutions/nb01_ex02.py
@@ -1,29 +1,20 @@
 import pandas as pd
 
-def create_model_input_table(
-    shuttles: pd.DataFrame, companies: pd.DataFrame, reviews: pd.DataFrame
-) -> pd.DataFrame:
-    """Combines all data to create a model input table.
-
-    Args:
-        shuttles: Preprocessed data for shuttles.
-        companies: Preprocessed data for companies.
-        reviews: Raw data for reviews.
-    Returns:
-        Model input table.
-
-    """
-    rated_shuttles = shuttles.merge(reviews, left_on="id", right_on="shuttle_id")
-    rated_shuttles = rated_shuttles.drop("id", axis=1)
-    model_input_table = rated_shuttles.merge(
-        companies, left_on="company_id", right_on="id"
-    )
-    model_input_table = model_input_table.dropna()
-    return model_input_table
+def _parse_money(x: pd.Series) -> pd.Series:
+    x = x.str.replace("$", "", regex=False).str.replace(",", "", regex=False)
+    x = x.astype(float)
+    return x
 
 
-create_model_input_table_node = node(
-    func=create_model_input_table,
-    inputs=["shuttles", "companies", "reviews"],
-    outputs="model_input_table"
+def preprocess_shuttles(shuttles: pd.DataFrame) -> pd.DataFrame:
+    shuttles["d_check_complete"] = _is_true(shuttles["d_check_complete"])
+    shuttles["moon_clearance_complete"] = _is_true(shuttles["moon_clearance_complete"])
+    shuttles["price"] = _parse_money(shuttles["price"])
+    return shuttles
+
+
+preprocess_shuttles_node = node(
+    func=preprocess_shuttles,
+    inputs="shuttles",
+    outputs="preprocessed_shuttles"
 )
