+++ kedro-databricks-bootcamp/01_intro/solutions/nb01_ex03.py
@@ -1,5 +1,29 @@
-data_processing = pipeline([
-    preprocess_companies_node,
-    preprocess_shuttles_node,
-    create_model_input_table_node,
-])
+import pandas as pd
+
+def create_model_input_table(
+    shuttles: pd.DataFrame, companies: pd.DataFrame, reviews: pd.DataFrame
+) -> pd.DataFrame:
+    """Combines all data to create a model input table.
+
+    Args:
+        shuttles: Preprocessed data for shuttles.
+        companies: Preprocessed data for companies.
+        reviews: Raw data for reviews.
+    Returns:
+        Model input table.
+
+    """
+    rated_shuttles = shuttles.merge(reviews, left_on="id", right_on="shuttle_id")
+    rated_shuttles = rated_shuttles.drop("id", axis=1)
+    model_input_table = rated_shuttles.merge(
+        companies, left_on="company_id", right_on="id"
+    )
+    model_input_table = model_input_table.dropna()
+    return model_input_table
+
+
+create_model_input_table_node = node(
+    func=create_model_input_table,
+    inputs=["shuttles", "companies", "reviews"],
+    outputs="model_input_table"
+)
