+++ kedro-databricks-bootcamp/01_intro/solutions/nb01_ex04.py
@@ -1,0 +1,5 @@
+data_processing = pipeline([
+    preprocess_companies_node,
+    preprocess_shuttles_node,
+    create_model_input_table_node,
+])
